moved all of the utility functions to the new dnpMath; *TODO* t2_function not yet fixed
